Added OrderController

+++ komis-backend/CarShop/src/main/java/com/TAB/CarShop/Repositories/DealerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DealerRepository extends JpaRepository<Dealer, Integer> {
+public interface DealerRepository extends JpaRepository<Dealer, Long> {
 }
